+++ src/RcppExports.cpp
@@ -331,6 +331,47 @@ RcppExport SEXP _remstats_degree_tie(SEXP typeSEXP, SEXP edgelistSEXP, SEXP adjm
     UNPROTECT(1);
     return rcpp_result_gen;
 }
+// degree_undirected_tie
+arma::mat degree_undirected_tie(int type, const arma::mat& edgelist, const arma::mat& adjmat, const arma::vec& actors, const arma::vec& types, int start, int stop, bool consider_type);
+static SEXP _remstats_degree_undirected_tie_try(SEXP typeSEXP, SEXP edgelistSEXP, SEXP adjmatSEXP, SEXP actorsSEXP, SEXP typesSEXP, SEXP startSEXP, SEXP stopSEXP, SEXP consider_typeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< int >::type type(typeSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type edgelist(edgelistSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type adjmat(adjmatSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type actors(actorsSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type types(typesSEXP);
+    Rcpp::traits::input_parameter< int >::type start(startSEXP);
+    Rcpp::traits::input_parameter< int >::type stop(stopSEXP);
+    Rcpp::traits::input_parameter< bool >::type consider_type(consider_typeSEXP);
+    rcpp_result_gen = Rcpp::wrap(degree_undirected_tie(type, edgelist, adjmat, actors, types, start, stop, consider_type));
+    return rcpp_result_gen;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _remstats_degree_undirected_tie(SEXP typeSEXP, SEXP edgelistSEXP, SEXP adjmatSEXP, SEXP actorsSEXP, SEXP typesSEXP, SEXP startSEXP, SEXP stopSEXP, SEXP consider_typeSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_remstats_degree_undirected_tie_try(typeSEXP, edgelistSEXP, adjmatSEXP, actorsSEXP, typesSEXP, startSEXP, stopSEXP, consider_typeSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error(CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
 // triad_tie
 arma::mat triad_tie(int type, const arma::mat& edgelist, const arma::vec& actors, const arma::vec& types, const arma::mat& adjmat, const arma::mat& riskset, int start, int stop, bool consider_type);
 static SEXP _remstats_triad_tie_try(SEXP typeSEXP, SEXP edgelistSEXP, SEXP actorsSEXP, SEXP typesSEXP, SEXP adjmatSEXP, SEXP risksetSEXP, SEXP startSEXP, SEXP stopSEXP, SEXP consider_typeSEXP) {
@@ -1148,6 +1189,7 @@ static int _remstats_RcppExport_validate(const char* sig) {
         signatures.insert("arma::mat(*inertia_tie)(const arma::mat&,const arma::mat&,const arma::mat&,int,bool,const arma::vec&,int,int,bool)");
         signatures.insert("arma::mat(*reciprocity_tie)(const arma::mat&,const arma::mat&,const arma::mat&,int,const arma::vec&,int,int,bool)");
         signatures.insert("arma::mat(*degree_tie)(int,const arma::mat&,const arma::mat&,const arma::vec&,const arma::vec&,int,int,bool)");
+        signatures.insert("arma::mat(*degree_undirected_tie)(int,const arma::mat&,const arma::mat&,const arma::vec&,const arma::vec&,int,int,bool)");
         signatures.insert("arma::mat(*triad_tie)(int,const arma::mat&,const arma::vec&,const arma::vec&,const arma::mat&,const arma::mat&,int,int,bool)");
         signatures.insert("arma::mat(*pshift_tie)(int,const arma::mat&,const arma::mat&,int,int,int,int,bool)");
         signatures.insert("arma::rowvec(*rankR)(arma::rowvec,int)");
@@ -1182,6 +1224,7 @@ RcppExport SEXP _remstats_RcppExport_registerCCallable() {
     R_RegisterCCallable("remstats", "_remstats_inertia_tie", (DL_FUNC)_remstats_inertia_tie_try);
     R_RegisterCCallable("remstats", "_remstats_reciprocity_tie", (DL_FUNC)_remstats_reciprocity_tie_try);
     R_RegisterCCallable("remstats", "_remstats_degree_tie", (DL_FUNC)_remstats_degree_tie_try);
+    R_RegisterCCallable("remstats", "_remstats_degree_undirected_tie", (DL_FUNC)_remstats_degree_undirected_tie_try);
     R_RegisterCCallable("remstats", "_remstats_triad_tie", (DL_FUNC)_remstats_triad_tie_try);
     R_RegisterCCallable("remstats", "_remstats_pshift_tie", (DL_FUNC)_remstats_pshift_tie_try);
     R_RegisterCCallable("remstats", "_remstats_rankR", (DL_FUNC)_remstats_rankR_try);
@@ -1215,6 +1258,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_remstats_inertia_tie", (DL_FUNC) &_remstats_inertia_tie, 9},
     {"_remstats_reciprocity_tie", (DL_FUNC) &_remstats_reciprocity_tie, 8},
     {"_remstats_degree_tie", (DL_FUNC) &_remstats_degree_tie, 8},
+    {"_remstats_degree_undirected_tie", (DL_FUNC) &_remstats_degree_undirected_tie, 8},
     {"_remstats_triad_tie", (DL_FUNC) &_remstats_triad_tie, 9},
     {"_remstats_pshift_tie", (DL_FUNC) &_remstats_pshift_tie, 8},
     {"_remstats_rankR", (DL_FUNC) &_remstats_rankR, 2},